requirements.txt
click==8.1.7
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.4
MarkupSafe==2.1.5
Werkzeug==2.3.7
gunicorn==20.1.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
Flask-Limiter==3.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
python-dotenv==1.0.0
email-validator==2.0.0
supabase==1.0.1
httpx==0.23.3